requirements.txt
requests==2.32.0
reportportal-client==5.5.4
pytest==8.1.1
selenium==4.12.0
python_dotenv==0.16.0
Appium_Python_Client==3.2.0
pytest-xdist>=1.31
pytest-html>=3.0.0
pytest-rerunfailures>=9.1.1
pytest_reportportal==5.4.0
pillow>=6.2.0
tesults==1.2.1
boto3==1.33.0
loguru
imageio
questionary>=1.9.0
clear-screen>=0.1.14
prompt-toolkit==2.0.10
axe_selenium_python==2.1.6
pytest-snapshot==0.9.0
beautifulsoup4>=4.12.3
openai==1.12.0
jinja2==3.1.3
pytesseract==0.3.10
pytest-asyncio==0.23.7
prettytable==3.10.2
setuptools>=75.1.0; python_version >= '3.12'
